move bech32_mod under blsct directory
