+++ src/Makefile.test.include
@@ -77,7 +77,6 @@ BITCOIN_TESTS =\
   test/base58_tests.cpp \
   test/base64_tests.cpp \
   test/bech32_tests.cpp \
-  test/bech32_mod_tests.cpp \
   test/bip32_tests.cpp \
   test/blockchain_tests.cpp \
   test/blockencodings_tests.cpp \
@@ -89,6 +88,7 @@ BITCOIN_TESTS =\
   test/blsct/arith/mcl/mcl_g1point_tests.cpp \
   test/blsct/arith/mcl/mcl_integration_tests.cpp \
   test/blsct/arith/mcl/mcl_scalar_tests.cpp \
+  test/blsct/bech32_mod_tests.cpp \
   test/blsct/building_block/generator_deriver_tests.cpp \
   test/blsct/building_block/imp_inner_prod_arg_tests.cpp \
   test/blsct/building_block/lazy_points_tests.cpp \
