+++ src/blsct/bech32_mod.cpp
@@ -4,7 +4,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include <blsct/double_public_key.h>
-#include <bech32_mod.h>
+#include <blsct/bech32_mod.h>
 #include <util/vector.h>
 
 #include <array>
