+++ src/test/blsct/bech32_mod_tests.cpp
@@ -3,8 +3,8 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include <blsct/bech32_mod.h>
 #include <blsct/double_public_key.h>
-#include <bech32_mod.h>
 #include <test/util/str.h>
 #include <util/strencodings.h>
