+++ src/test/fuzz/bech32_mod.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <bech32_mod.h>
+#include <blsct/bech32_mod.h>
 #include <test/fuzz/fuzz.h>
 #include <test/util/str.h>
 #include <util/strencodings.h>
